refactor: combine file import and geocode tests per review feedback
